Made a toString method for recordings to test whether they are functioning correctly,
created a test main in recorder to test creation of a recording, saving and loading.

Created a save recording method in Recorder which
saves a recording object as a JSON file in the recordings folder,

created the loadRecording method in replay which loads a Recording object from a JSON file in the recordings folder.

altered constructor For the replay class to take a String that is the fileName of the recording it will be replaying.